+++ bigsort/V2/InfixComparer.cs
@@ -8,12 +8,12 @@ namespace BigSort.V2
   /// <summary>
   /// Sorts infixes in the alphabetical order.
   /// </summary>
-  internal class InfixComparer : IComparer<long>
+  internal class InfixComparer : IComparer<uint>
   {
     /// <summary>
     /// IComparer.
     /// </summary>
-    public int Compare([AllowNull] long x, [AllowNull] long y)
+    public int Compare([AllowNull] uint x, [AllowNull] uint y)
     {
       var infix1 = Encoding.Unicode.GetString(BitConverter.GetBytes(x));
       var infix2 = Encoding.Unicode.GetString(BitConverter.GetBytes(y));
